src/requirements.txt
# Requirements for Data Mining Pipeline
requests
PyGithub
python-dotenv
tqdm
pandas